// src/platform/packages/shared/esql/kbn-esql-language/src/commands/registry/types.d.ts
export interface ESQLUserDefinedColumn {
  name: string;
  type: SupportedDataType | 'unknown';
  userDefined: true;
  location: ESQLLocation;
  isUnmappedField?: boolean;
}
export type ESQLColumnData = ESQLUserDefinedColumn | ESQLFieldWithMetadata;
// ...
export declare enum UnmappedFieldsStrategy {
  DEFAULT = 'DEFAULT',
  NULLIFY = 'NULLIFY',
  LOAD = 'LOAD',
  LOAD_ALL = 'LOAD_ALL',
}